requirements.txt
streamlit
numpy
opencv-python
pillow
inference-sdk